Fix updaters tests
Scripts do not depend on index anymore
