+++ back/prisma/scripts/set-company-name-gerepid-naf.ts
@@ -1,4 +1,4 @@
-import { Updater, registerUpdater } from ".";
+import { Updater, registerUpdater } from "./helper/helper";
 import { prisma } from "../../src/generated/prisma-client";
 import { searchCompany } from "../../src/companies/sirene";
